modified files in 'www' folder

+++ www/boothInfo_detail.php
@@ -38,25 +38,6 @@ $row = $rst->fetch_assoc();
 									</td>
 								</tr>
 
-								<?php
-								if($_SESSION['level']=='admin') {
-								?>
-								<tr>
-									<td>
-										<strong>ApLevel </strong>&nbsp;&nbsp;<?php echo $row['apLevel'];?>
-									</td>
-									<td >
-										<strong>NFCTagID </strong>&nbsp;&nbsp;<?php echo $row['nfcTagId'];?>
-									</td>
-								</tr>
-
-								<?php 
-								} else if($_SESSION['level']=='user') {
-
-								}
-								?>
-
-
 
 									<tr><td colspan="2"><strong>브로셔이미지 </strong></td>
 									</tr>
@@ -69,16 +50,25 @@ $row = $rst->fetch_assoc();
 								</tbody>
 							</table>
 							</div>
-							
+
 								<div class="col-lg-2">
 								<input type="button" class="btn btn-primary btn-sm" value="수정" onclick="location.href='boothInfo_post.php?mode=correct&no=<?php echo $row['index']?>';"/>
+							<?
+							if($_SESSION['IDlevel']=='admin') {
+							?>	
+
 								<input type="button" class="btn btn-primary btn-sm" value="삭제" onclick="deleteBoard();"/>
+							
+
 
 								</div>
 								<div class="col-lg-9"></div>
 								<div class="col-lg-1">
 								<input type="button" class="btn btn-primary btn-sm" onclick="location.href='board.php?type=BoothInfo';" value="돌아가기" />
 								</div>
+								<?
+							}
+							?>
 							</div>
 							</div>
